Remove pack/build distinction
Might not be needed
